+++ anomsmith/primitives/detectors/__init__.py
@@ -1,15 +1,24 @@
 """Anomaly detectors."""
 
+from anomsmith.primitives.detectors.autoencoders import (
+    LSTMAutoencoderDetector,
+    PyTorchAutoencoderDetector,
+)
 from anomsmith.primitives.detectors.change_point import ChangePointDetector
 from anomsmith.primitives.detectors.drift import ARIMADriftDetector
-from anomsmith.primitives.detectors.ensemble import VotingEnsembleDetector
+from anomsmith.primitives.detectors.ensemble import (
+    EnsembleDetector,
+    ScoreCombiningEnsembleDetector,
+    VotingEnsemble,
+    VotingEnsembleDetector,
+)
 from anomsmith.primitives.detectors.ml import (
     IsolationForestDetector,
     LOFDetector,
     RobustCovarianceDetector,
 )
 from anomsmith.primitives.detectors.pca import PCADetector
-from anomsmith.primitives.detectors.wavelet import WaveletDetector
+from anomsmith.primitives.detectors.wavelet import WaveletDenoiser, WaveletDetector
 
 __all__ = [
     "ChangePointDetector",
@@ -18,6 +27,12 @@ __all__ = [
     "RobustCovarianceDetector",
     "PCADetector",
     "WaveletDetector",
+    "WaveletDenoiser",
     "VotingEnsembleDetector",
+    "VotingEnsemble",
+    "ScoreCombiningEnsembleDetector",
+    "EnsembleDetector",
+    "LSTMAutoencoderDetector",
+    "PyTorchAutoencoderDetector",
     "ARIMADriftDetector",
 ]